SMILE-852-kaltura-list - translations, oppretting av ressurs ved valg
