+++ src/translations/translations_NO.json
@@ -218,8 +218,10 @@
     "uib": "UiB"
   },
   "kaltura": {
+    "choose_a_resource": "Velg en ressurs fra listen",
     "my_resources" : "Mine ressurser i Kaltura",
-    "choose_a_resource": "Velg en ressurs fra listen"
+    "show_my_resources" : "Vis mine ressurser fra Kaltura",
+    "start_with_kaltura_resource" : "Start med en ressurs fra Kaltura"
   },
   "license": {
     "commercial_options": {
